Add reset to QR screen and remove reset from auto (#47)
* Add reset to QR screen and remove reset from auto

* fix flutter analyze issues
